Fix tool-default.svg icon usage (#628)
This fixes an issue where rendering multiple instances of the tool-default.svg icon at once can cause filter effects to cascade to instances where they shouldn't be applied. This is because they all have the same underlying filter ID. 

Additionally, this renders the symbol in white when in dark mode. 

Fixes #618