Ensure meta/links is loaded correctly/dynamically (#644)
This should finally fix the issues that I created implementing #639.
`<meta />` and `<links />` data is, for the most part, passed passed
through the loader. This is necessary for `<link rel="canonical" />`
because it requires a complete URL, and so we need to know the `ORIGIN`
which we get from the server. The meta tag _could_ be loaded without
using the server in most cases, but that requires splitting out the
`seo.server.ts` functions, and this feels more consistent anyway.

Also removes the `/sdks` page which doesn't exist and is being
redirected anyway.