Ensure extension stripped from canonical URL (#648)
The previous change used `path` directory from the `params` to determine
the canonical URL, but that may include an extension which should be
stripped. This ensures that extension is no longer included.

This also fixes an issue with the `stripExtension` regex which wasn't
previously working correctly with `.mdx` extensions because it would
catch the `.md` part of the expression first, leaving a single "x"
behind. (i.e. "foo.mdx" would become "foox")