Generate preview links checks for github (#663)
Closes #655 

Generates a list of preview links for a Github branch/PR when changes
are made.

This implementation uses the Github "Checks" API. When any changed is
pushed, Github sends us a `check_suite` event. We respond by initiating
a `check_run` which

This is the same workflow used for things like linting checks or unit
tests. In this case we just always provide a "neutral" status as opposed
to a `failure` or `success` which would typical of something like a
linter. The summary we provide includes the preview links.

This also allows a user to trigger a "re-run" on-demand, if needed.

## Screenshot

The result looks something like this (though this is using a temporary
test Github App I created):

<img width="971" alt="Screen Shot 2022-09-21 at 7 36 48 PM"
src="https://user-images.githubusercontent.com/393220/191758412-817b8722-0cd9-4668-bb41-e572923cf4bb.png">

## Code

Github requires that check_run/check_suite events authenticate as a
GitHub App, which meant that simply using a token wasn't going to work.
So now we instantiate a Github App instance which handles the
authentication for us. It also has an eventing system for responding to
webhooks automatically (and everything is typed correctly, which is
great).

So instead of manually checking for specific events ourselves we can
just instantiate the app and subscribe to events:

```
app.on("push", (pushEvent) => {
   // handle push webhook
});

app.on("check_suite", (checkSuiteEvent) => {
   // handle check_suite webhook
});
```


And in our router we just tell the app to receive the event:

```
app.webhooks.receive({ id, name, payload: body });
```

So that cleans things up quite a bit.

I also started reorganizing some things in the `/cms` folder to
consolidate, for example. github-specific code.