Check for redirects in catch-all route (#713)
Closes #704 

The current redirect is in the loader of the root layout, and this
doesn't get "refreshed" when we change location client-side. The
catch-all _does_ get rerendered on every page change, so if we redirect
here it works as expected.

It feels like there should be a better, "global" way to do this sort of
thing, but I wasn't able to find anything in the remix docs that would
make this possible/simpler.